pyproject.toml
[tool.poetry]
name = "git-settings-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
PyGithub = "^2.1.1"
fastapi = {extras = ["uvicorn"], version = "^0.104.1"}
pydantic = {extras = ["email"], version = "^2.5.2"}
pydantic-settings = "^2.1.0"
PyYAML = "^6.0.1"
uvicorn = "^0.24.0.post1"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
